Cleaning int przed czyszczeniem
### CHANGED
- Masowa konwersja typów numerycznych: Zmieniono typy bazowe wszystkich zmiennych z `float` / `double` na `int` dla kosztów, produkcji oraz powiązanych liczników w całej bazie kodowej. Zmiana ta wynika z faktu, że w praktyce gra korzysta wyłącznie z wartości całkowitych.
- Aktualizacja klas i struktur: Zmiany objęły składowe klas, konstruktory oraz gettery w modułach: `Building`, `Energy`, `Farm`, `Housing`, `Logistics`, `Colony`, `Enums` oraz `Game`.
- Sygnatury funkcji: Zmodyfikowano wirtualną metodę `work()`, aby zwracała typ `int`. Zaktualizowano również struktury wynikowe (`BuildResult`, `DestroyResult`, `NextResult`), wprowadzając do nich pola całkowite.
- Przetwarzanie zasobów: Dostosowano mechanikę zapisu/odczytu stanu gry, logikę nowej tury oraz systemy budowania i wyburzania do semantyki liczb całkowitych.
- Poprawki w UI (Graphics): Zaktualizowano formatowanie w ImGui (zamiana formatowania na `%d`), dostosowano operacje matematyczne oraz uporządkowano przepływ UI (poprawne pary `PushID`/`PopID` i przełączniki kategorii budowania).

### FIXED
- Logika dzielenia przy zwrotach: Poprawiono obsługę dzielenia całkowitego przy odzyskiwaniu surowców ze zburzonych budynków – zaimplementowano zaokrąglanie w górę (np. operacja $5 / 2$ zwraca teraz poprawnie $3$).
- Błąd logiczny operatora: Naprawiono błąd składniowy w warunku logicznym, zamieniając błędny operator bitowy na logiczny (`&` -> `&&`).

+++ src/building.cpp
@@ -14,7 +14,7 @@ int Building::next_id=1;
 // ==========================================
 
 Building:: Building():kosztEnergii(0),kosztKamien(0),kosztTytan(0),type(TypBudynku::NIEZNANY),name("Budynek"),id(next_id++),workers(0),residents(0){}
-Building:: Building(string n,TypBudynku t, double kE,double kK, double kT,int w):kosztEnergii(kE),kosztKamien(kK),kosztTytan(kT),type(t),name(n),id(next_id++),workers(w), residents(0){}
+Building:: Building(string n,TypBudynku t, int kE, int kK, int kT,int w):kosztEnergii(kE),kosztKamien(kK),kosztTytan(kT),type(t),name(n),id(next_id++),workers(w), residents(0){}
 
 // ==========================================
 // VIRTUALNE
@@ -26,7 +26,7 @@ void Building:: prnt(int il)const{}
 void Building:: UIprnt(int il)const{}
 
 //WYKONYWANIE PRACY PRZEZ DANY BUDYNEK PODCZAS nextRound
-double Building:: work(){return 0;}
+int Building:: work(){return 0;}
 
 //ZAPISYWANIE DANYCH O BUDYNKU DO PLIKU
 void Building::save(ofstream& plik)const{
@@ -53,6 +53,6 @@ string Building:: getName()const {return name;}
 TypBudynku Building::getTyp() const {return type;}
 int Building::getResidents() const{return residents;}
 int Building::getDemandWorkers()const {return workers;}
-double Building::getReqEnergy() const{return kosztEnergii;}
-double Building::getKosztKamien() const{return kosztKamien;}
-double Building::getKosztTytan() const{return kosztTytan;}
+int Building::getReqEnergy() const{return kosztEnergii;}
+int Building::getKosztKamien() const{return kosztKamien;}
+int Building::getKosztTytan() const{return kosztTytan;}
